fix undefined behavior in getBlockHeader

+++ lib/codegen/CreateTerm.cpp
@@ -131,8 +131,8 @@ llvm::StructType *getBlockType(llvm::Module *Module, const KOREObjectSymbol *sym
 llvm::Value *getBlockHeader(llvm::Module *Module, const KOREObjectSymbol *symbol) {
   llvm::StructType *BlockHeaderType = Module->getTypeByName(BLOCKHEADER_STRUCT);
   uint64_t headerVal = symbol->getTag();
-  headerVal |= symbol->length() << 32;
-  headerVal |= symbol->getLayout() << 48;
+  headerVal |= (uint64_t)symbol->length() << 32;
+  headerVal |= (uint64_t)symbol->getLayout() << 48;
   return llvm::ConstantStruct::get(BlockHeaderType, llvm::ConstantInt::get(llvm::IntegerType::get(Module->getContext(), 64), headerVal));
 }
